feat(#6): upgrade Astro v5 to v6
- Migrate content collections to Content Layer API (glob loader)
- Move config to src/content.config.ts (required by v6)
- Await OGImageRoute (now async in astro-og-canvas v0.11.1)
- Add @astrojs/internal-helpers as direct dep to fix pnpm hoisting
- Update astro, @astrojs/mdx, astro-og-canvas, rss, sitemap
