+++ package.json
@@ -10,13 +10,14 @@
     "generate-history": "tsx scripts/generate-git-history.ts"
   },
   "dependencies": {
-    "@astrojs/mdx": "^4.3.13",
-    "@astrojs/rss": "^4.0.15",
-    "@astrojs/sitemap": "^3.7.0",
+    "@astrojs/internal-helpers": "0.9.0",
+    "@astrojs/mdx": "^5.0.4",
+    "@astrojs/rss": "^4.0.18",
+    "@astrojs/sitemap": "^3.7.2",
     "@fontsource/geist-mono": "^5.2.7",
     "@fontsource/geist-sans": "^5.2.5",
-    "astro": "^5.17.2",
-    "astro-og-canvas": "^0.7.2",
+    "astro": "^6.3.0",
+    "astro-og-canvas": "^0.11.1",
     "canvaskit-wasm": "^0.40.0",
     "diff": "^8.0.3",
     "geist": "1.7.0",
