+++ src/pages/api/og-image/[default].ts
@@ -1,6 +1,6 @@
 import { OGImageRoute } from 'astro-og-canvas';
 
-export const { getStaticPaths, GET } = OGImageRoute({
+export const { getStaticPaths, GET } = await OGImageRoute({
   param: 'default',
   pages: {
     'default': {
